+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - [🚀] `BackGestureDetector`增加`width`参数;
 - [🚀] 日志增加所在类名和方法名;
 - [🚀] `ListX`增加removeType扩展;
+- [🚀] `ImageView`增加`rotation`参数;
+- [🚀] `BuildContext`增加`getPageStorage`和`setPageStorage`扩展;
 
 ## 0.50.4
 - [🚀] 增加`CoordConvert`, 用于坐标系转换;